Started modularizing the Auth Modal. Switching machines
