+++ client/src/components/dom/Auth/Auth.module.scss
@@ -18,4 +18,10 @@
     color: white;
     font-size: 1.2em;
   }
+
+  p {
+    padding: 1em 2em;
+    color: white;
+    font-size: 1.2em;
+  }
 }
