+++ client/src/components/dom/Auth/AuthModal/RegisterModal/index.ts
@@ -1,0 +1,3 @@
+import RegisterModal from "./RegisterModal";
+
+export default RegisterModal;
